Rename soil-optimistic package to soil.plant.compose.optimistic
This change aligns the optimistic update package with our naming convention, following the same pattern as
`soil-lazyload` which uses the `soil.plant` prefix.

Since the `soil-optimistic` package has not been publicly released yet, there is no impact on existing dependencies or users.
